+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 
 - python: fix crash on quit with Python 3.12 ([#2046](https://github.com/weechat/weechat/issues/2046), [#2126](https://github.com/weechat/weechat/issues/2126))
 - core: fix crash when deleting a bar that has no items ([#2138](https://github.com/weechat/weechat/issues/2138))
+- relay/api: fix crash when sending data to a remote buffer when the remote has been deleted ([#2157](https://github.com/weechat/weechat/issues/2157))
 - irc, xfer: fix display of input prompt in IRC private buffers and DCC chat buffers ([#2128](https://github.com/weechat/weechat/issues/2128))
 - core: apply buffer properties (options weechat.buffer.*) when a buffer name is changed
 - irc: fix property "short_name" of channel buffer when the joined channel has a different case than the `/join` command